[CD-192] Move: make model folder
