+++ ai/sfen/model/generator.py
